HSEARCH-3115 Move classes related to Elasticsearch index settings to a dedicated package
