+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/index/settings/impl/esnative/IndexSettings.java
@@ -1,0 +1,37 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.backend.elasticsearch.index.settings.impl.esnative;
+
+import com.google.gson.GsonBuilder;
+
+/**
+ * Settings for an Elasticsearch index.
+ *
+ * @author Yoann Rodiere
+ */
+public class IndexSettings {
+
+	private Analysis analysis;
+
+	public Analysis getAnalysis() {
+		return analysis;
+	}
+
+	public void setAnalysis(Analysis analysis) {
+		this.analysis = analysis;
+	}
+
+	public boolean isEmpty() {
+		return analysis == null || analysis.isEmpty();
+	}
+
+	@Override
+	public String toString() {
+		return new GsonBuilder().setPrettyPrinting().create().toJson( this );
+	}
+
+}
